Adds minimal support for Array of Tables (AoT)
Inline tables are a cool feature of the TOML, and very useful as
elements of an array.

However there are some circumstances that they are explicitly
discouraged by spec itself.

For example when dumping a list of dicts that in turn
have lists inside of them, `tomli-w` will introduce line breaks
inside of inline tables, which is not recommended by the spec

This situation was first described in issue #12.

Please notice this usage pattern is not that uncommon. We can find
examples of this nesting in the wild, like the ones used in `mypy`
configuration:

https://mypy.readthedocs.io/en/stable/config_file.html#using-a-pyproject-toml-file